+++ crates/api/src/neighbor_service.rs
@@ -164,6 +164,7 @@ impl proto::neighbor_service_server::NeighborService for NeighborService {
             gr_stale_routes_time: 360,
             local_ipv6_nexthop: None,
             route_reflector_client: false,
+            route_server_client: false,
             add_path_receive: false,
             add_path_send: false,
             add_path_send_max: 0,