+++ crates/api/src/peer_types.rs
@@ -60,6 +60,7 @@ pub struct PeerManagerNeighborConfig {
     pub gr_stale_routes_time: u64,
     pub local_ipv6_nexthop: Option<Ipv6Addr>,
     pub route_reflector_client: bool,
+    pub route_server_client: bool,
     pub add_path_receive: bool,
     pub add_path_send: bool,
     pub add_path_send_max: u32,